CVE Service & Mary installed
I have constructed a CVE monitoring and notification service - CVEs can be added to applications, a job (scheduled daily) will run the registered queries and examine the results.

If a CVE listing has a severity rating equal to our higher than the severity threshold for that query - the CVE text is sent via notification pipeline to the email addresses specified in the query.

Mary UI has been installed for use in the front-end.